requirement.txt
google-ai-generativelanguage==0.6.6
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.63.2
numpy==1.26.4
opencv-python==4.10.0.84
pandas==2.2.2
pillow==10.4.0
pytesseract==0.3.13
python-dotenv==1.0.1